given_data/requirements.txt
amqp==5.2.0
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
kombu==5.3.6
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
pandas==2.2.1
prompt-toolkit==3.0.43
PyMySQL==1.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.29
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
